+++ src/main/java/calculator/OutputHandler.java
@@ -1,7 +1,9 @@
 package calculator;
 
+import java.math.BigInteger;
+
 public class OutputHandler {
-    public static void printOutput(int result) {
+    public static void printOutput(BigInteger result) {
         System.out.println("결과 : " + result);
     }
 }
